Realisation de l'us 245

+++ src/app/ArbreCategorie/arbreCategorie.component.ts
@@ -5,8 +5,8 @@ import {CategorieNode} from "../../../e-commerce-ui-common/models/CategorieNode"
 import {CategorieFlatNode} from "../../../e-commerce-ui-common/models/CategorieFlatNode";
 import {ArbreService} from "../../../e-commerce-ui-common/business/arbre.service";
 import {Categorie} from "../../../e-commerce-ui-common/models/Categorie";
-import {CategorieBusinessService} from "../../../e-commerce-ui-common/business/categorie.service";
-import {element} from "protractor";
+import {Modal} from "ngx-modialog/plugins/bootstrap";
+
 
 /**
  * @title Arbre avec FlatNode
@@ -22,11 +22,16 @@ export class ArbreCategorieComponent implements OnInit {
   ngOnInit(): void {
   }
 
+
+
+
+
+
   public treeControl: FlatTreeControl<CategorieFlatNode>;
   public treeFlattener: MatTreeFlattener<CategorieNode, CategorieFlatNode>;
   public dataSource: MatTreeFlatDataSource<CategorieNode, CategorieFlatNode>;
 
-  constructor(private arbreService: ArbreService, categorieBusiness: CategorieBusinessService) {
+  constructor(private arbreService: ArbreService, private modal:Modal) {
     this.treeFlattener = new MatTreeFlattener(arbreService.transformerNodeToFlatNode, arbreService.getLevel,
       arbreService.isExpandable, arbreService.getChildren);
     this.treeControl = new FlatTreeControl<CategorieFlatNode>(arbreService.getLevel, arbreService.isExpandable);
@@ -104,24 +109,71 @@ export class ArbreCategorieComponent implements OnInit {
     node.enableToolNode = true;
   }
 
+
+  /**
+   * Methode permettant de supprimer une categorie a partir de sa flat node
+   * @param {CategorieFlatNode} node flat node représentant la categorie a supprimer
+   * @return {Promise<void>}
+   */
   async suprimerCategorie(node: CategorieFlatNode) {
-    //On appelle la methode supprimerCategorie du service categorieBusiness en passant par arbreService
-    let retourAPI = await this.arbreService.categorieBusiness.supprimerCategorie(new Categorie(node.id, node.nomCategorie, null, null));
-    if (retourAPI != null && retourAPI != undefined) {
-      //On met à jour le nom de la categorie afficher dans la node concerné
-      if (retourAPI) {
-        node.justDeleted=true;
-        node.nomCategorie="La categorie a bien était supprimé";
-      }
-      else {
-        console.log("pas supprimer");
-      }
-    } else{
-      //TODO Message erreur
-    }
+  //Modal
+    const dialogRef = this.modal.confirm()
+      .size('lg')
+      .isBlocking(true)
+      .showClose(false)
+      .keyboard(27)
+      .title('Suppresion de la catégorie ' + node.nomCategorie + ' - id(' + node.id + ')')
+      .body('Comfirmez vous la supression de la categorie ' + node.nomCategorie + ' - id(' + node.id + ')?')
+      .okBtn('Comfirmer la suppression')
+      .okBtnClass('btn btn-danger')
+      .cancelBtn('Annuler la supression')
+      .open();
+    dialogRef.result
+      .then(async() => {
+        //On appelle la methode supprimerCategorie du service categorieBusiness en passant par arbreService
+        let retourAPI = await this.arbreService.categorieBusiness.supprimerCategorie(new Categorie(node.id, node.nomCategorie, null, null));
+        if (retourAPI != null && retourAPI != undefined) {
+          //On met à jour le nom de la categorie afficher dans la node concerné
+          if (retourAPI) {
+            // On supprime la categorie visuelement si la suppression dans la base de donnée est un succès
+            this.deleteNode(node);
+          }
+          else {
+            console.log("pas supprimer");
+          }
+        } else{
+          //TODO Message erreur
+        }
+      }).catch(() => null); // Pour éviter l'erreur de promise dans console.log
+
   }
 
 
+  /**
+   * Methode permettant de supprimer une categorie visuelement
+   * @param {CategorieFlatNode} node la flat node representant la categorie a supprimer
+   */
+  deleteNode(node: CategorieFlatNode) {
+    // Si la flat node possède un parent on la supprimer des enfants de ce parent
+    if(node.idParent!=undefined){
+      let nodeParent =undefined;
+      let nodeEnfant=this.arbreService.flatNodeMap.get(node);
+      this.arbreService.flatNodeMap.forEach(function (value) {
+        if(value.id === node.idParent){
+          nodeParent = value;
+        }
+      });
+      this.arbreService.deleteChild(nodeParent!, nodeEnfant);
+    }else{
+      // Si la flat node ne possède pas de parent on recharge l'arbre pour la supprimer de l'affichage
+      this.arbreService.initialize();
+    }
+
+  }
+
+  public hasCategories(){
+    return this.arbreService.hasCategories;
+  }
 
 
 }
